+++ src/main/java/org/jabref/gui/preferences/PreferencesDialogViewModel.java
@@ -19,6 +19,7 @@ import org.jabref.gui.preferences.entryeditor.EntryEditorTab;
 import org.jabref.gui.preferences.entryeditortabs.CustomEditorFieldsTab;
 import org.jabref.gui.preferences.exporter.ExportCustomizationTab;
 import org.jabref.gui.preferences.external.ExternalTab;
+import org.jabref.gui.preferences.fetcher.FetcherTab;
 import org.jabref.gui.preferences.file.FileTab;
 import org.jabref.gui.preferences.general.GeneralTab;
 import org.jabref.gui.preferences.groups.GroupsTab;