+++ src/main/java/org/jabref/logic/importer/FetcherApiPreferences.java
@@ -1,0 +1,13 @@
+package org.jabref.logic.importer;
+
+public class FetcherApiPreferences {
+    private final String worldcatKey;
+
+    public FetcherApiPreferences(String worldcatKey) {
+        this.worldcatKey = worldcatKey;
+    }
+
+    public String getWorldcatKey() {
+        return worldcatKey;
+    }
+}